DataExtractor_Graph/requirements.txt
fastapi[standard]
python-multipart
pillow
python-dotenv
pydantic
typing-extensions
slowapi
mistralai
pyyaml
python-jose[cryptography]
uvicorn[standard]
sqlalchemy
passlib[bcrypt]